chore (refactoring): Move dataupdatecoordinator into own file
