+++ custom_components/bermuda/device_tracker.py
@@ -14,9 +14,9 @@ from homeassistant.core import callback
 from homeassistant.helpers.dispatcher import async_dispatcher_connect
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from . import BermudaDataUpdateCoordinator
 from .const import DOMAIN
 from .const import SIGNAL_DEVICE_NEW
+from .coordinator import BermudaDataUpdateCoordinator
 from .entity import BermudaEntity
 
 _LOGGER = logging.getLogger(__name__)
